+++ test/CMakeLists.txt
@@ -22,6 +22,7 @@ if (ENABLE_FFT)
 endif()
 if (ENABLE_SOLVERS)
     add_subdirectory (solver)
+    add_subdirectory (solver/fem)
 endif()
 add_subdirectory (particle)
 add_subdirectory (region)
