+++ test/solver/fem/TestPeriodicBC_sin.cpp
@@ -1,5 +1,16 @@
-// dirichlet BCs scaling (5 times, 256^3)
-// sin(pi x)
+// Tests the FEM Poisson solver by solving the problem:
+//
+// -Laplacian(u) = pi^2 * sin(pi * x), x in [-1,1]
+// u(-1) = u(1) = 0
+//
+// Exact solution is u(x) = sin(pi * x)
+//
+// Here we use periodic BCs, so this should work
+// for other domains too as long as the domain 
+// length is the size of the period i.e. 2.
+//
+// Usage:
+//     ./TestPeriodicBC_sin <dim> --info 5
 
 #include "Ippl.h"
 
@@ -34,7 +45,6 @@ struct AnalyticSol {
 template <typename T, unsigned Dim>
 void testFEMSolver(const unsigned& numNodesPerDim, const T& domain_start = 0.0,
                    const T& domain_end = 1.0) {
-    // std::function<T(ippl::Vector<T, Dim> x)> f_sol,
     // start the timer
     static IpplTimings::TimerRef initTimer = IpplTimings::getTimer("initTest");
     IpplTimings::startTimer(initTimer);
@@ -62,12 +72,15 @@ void testFEMSolver(const unsigned& numNodesPerDim, const T& domain_start = 0.0,
 
     ippl::FieldLayout<Dim> layout(MPI_COMM_WORLD, domain, isParallel);
     Field_t lhs(mesh, layout, numGhosts);  // left hand side (updated in the algorithm)
-    Field_t rhs(mesh, layout, numGhosts);  // right hand side (set once)
+    Field_t rhs(mesh, layout, numGhosts);  // left hand side (updated in the algorithm)
+
+    Field_t analytical(mesh, layout, numGhosts);  // right hand side (set once)
+    auto view_analytical = analytical.getView();
 
     // Define boundary conditions
     BConds_t bcField;
     for (unsigned int i = 0; i < 2 * Dim; ++i) {
-        bcField[i] = std::make_shared<ippl::ZeroFace<Field_t>>(i);
+        bcField[i] = std::make_shared<ippl::PeriodicFace<Field_t>>(i);
     }
     lhs.setFieldBC(bcField);
     rhs.setFieldBC(bcField);
@@ -75,6 +88,8 @@ void testFEMSolver(const unsigned& numNodesPerDim, const T& domain_start = 0.0,
     // set rhs
     auto view_rhs = rhs.getView();
     auto ldom     = layout.getLocalNDIndex();
+
+    AnalyticSol<T, Dim> analytic;
 
     using index_array_type = typename ippl::RangePolicy<Dim>::index_array_type;
     ippl::parallel_for(
@@ -87,6 +102,7 @@ void testFEMSolver(const unsigned& numNodesPerDim, const T& domain_start = 0.0,
             const ippl::Vector<T, Dim> x = (iVec)*cellSpacing + origin;
 
             apply(view_rhs, args) = sinusoidalRHSFunction<T, Dim>(x);
+            apply(view_analytical, args) = analytic(x);
         });
 
     IpplTimings::stopTimer(initTimer);
@@ -108,12 +124,15 @@ void testFEMSolver(const unsigned& numNodesPerDim, const T& domain_start = 0.0,
     IpplTimings::startTimer(errorTimer);
 
     // Compute the error
-    AnalyticSol<T, Dim> analytic;
     const T relError = solver.getL2Error(analytic);
+
+    lhs = lhs - analytical;
+    T normError = norm(lhs) / norm(analytical);
 
     m << std::setw(10) << numNodesPerDim;
     m << std::setw(25) << std::setprecision(16) << cellSpacing[0];
     m << std::setw(25) << std::setprecision(16) << relError;
+    m << std::setw(25) << std::setprecision(16) << normError;
     m << std::setw(25) << std::setprecision(16) << solver.getResidue();
     m << std::setw(15) << std::setprecision(16) << solver.getIterationCount();
     m << endl;
@@ -128,6 +147,14 @@ int main(int argc, char* argv[]) {
 
         using T = double;
 
+        unsigned dim = 3;
+
+        if (argc > 1 && std::atoi(argv[1]) == 1) {
+            dim = 1;
+        } else if (argc > 1 && std::atoi(argv[1]) == 2) {
+            dim = 2;
+        }
+
         // start the timer
         static IpplTimings::TimerRef allTimer = IpplTimings::getTimer("allTimer");
         IpplTimings::startTimer(allTimer);
@@ -135,14 +162,26 @@ int main(int argc, char* argv[]) {
         msg << std::setw(10) << "Size";
         msg << std::setw(25) << "Spacing";
         msg << std::setw(25) << "Relative Error";
+        msg << std::setw(25) << "Norm Error";
         msg << std::setw(25) << "Residue";
         msg << std::setw(15) << "Iterations";
         msg << endl;
 
-        unsigned pt = std::atoi(argv[1]);
-
-        for (unsigned n = 0; n < 5; ++n) {
-            testFEMSolver<T, 3>(pt, -1.0, 1.0);
+        if (dim == 1) {
+            // 1D Sinusoidal
+            for (unsigned n = 1 << 3; n <= 1 << 10; n = n << 1) {
+                testFEMSolver<T, 1>(n, -1.5, 2.5);
+            }
+        } else if (dim == 2) {
+            // 2D Sinusoidal
+            for (unsigned n = 1 << 3; n <= 1 << 10; n = n << 1) {
+                testFEMSolver<T, 2>(n, -1.5, 2.5);
+            }
+        } else {
+            // 3D Sinusoidal
+            for (unsigned n = 1 << 3; n <= 1 << 9; n = n << 1) {
+                testFEMSolver<T, 3>(n, -1.5, 2.5);
+            }
         }
 
         // stop the timer
