+++ test/solver/fem/TestPeriodicBC_sinsin.cpp
@@ -1,41 +1,74 @@
-// Tests the FEM Poisson solver with a Gaussian source
-// and homogeneous Dirichlet boundaries
+// Tests the FEMPoissonSolver with periodic BCs
+// using the problem sin(sin(pi x)).
+// A convergence study is done by increasing the
+// problem size and looking at the relative error.
 //
 // Usage:
-//     ./TestFEMPoissonSolver_gaussian1d --info 5
+//      ./TestFEMPoissonSolver_periodic_sinsin --info 5
 
 #include "Ippl.h"
 
-#include "Meshes/Centering.h"
+#include <Kokkos_MathematicalConstants.hpp>
+#include <Kokkos_MathematicalFunctions.hpp>
+#include <cstdlib>
+#include <iostream>
+#include <string>
+#include <typeinfo>
+
+#include "Utility/Inform.h"
+#include "Utility/IpplTimings.h"
+
 #include "PoissonSolvers/FEMPoissonSolver.h"
 
-template <typename T>
-KOKKOS_INLINE_FUNCTION T gaussian1D(ippl::Vector<T, 1> x_vec, const T& sigma = 0.05, const T& mu = 0.5) {
+template <typename T, unsigned Dim>
+KOKKOS_INLINE_FUNCTION T sinusoidalRHSFunction(ippl::Vector<T, Dim> x_vec) {
     const T pi = Kokkos::numbers::pi_v<T>;
-    const T x  = x_vec[0];
 
-    const T prefactor = (1.0 / Kokkos::sqrt(2.0 * pi)) * (1.0 / sigma);
-    const T r2        = (x - mu) * (x - mu);
+    T val = 1.0;
+    if (Dim == 1) {
+        T x = x_vec[0];
 
-    return prefactor * Kokkos::exp(-r2 / (2.0 * sigma * sigma));
+        val = Kokkos::pow(pi, 2) * ((Kokkos::cos(Kokkos::sin(pi * x)) * Kokkos::sin(pi * x))
+                + (Kokkos::pow(Kokkos::cos(pi * x), 2) * Kokkos::sin(Kokkos::sin(pi * x))));
+
+    } else if (Dim == 2) {
+        T x = x_vec[0];
+        T y = x_vec[1];
+
+        val = Kokkos::pow(pi, 2)
+                * (Kokkos::cos(Kokkos::sin(pi * y)) * Kokkos::sin(pi * y) * Kokkos::sin(Kokkos::sin(pi * x))
+                + (Kokkos::cos(Kokkos::sin(pi * x)) * Kokkos::sin(pi * x)
+                + (Kokkos::pow(Kokkos::cos(pi * x), 2) + Kokkos::pow(Kokkos::cos(pi * y), 2)) * Kokkos::sin(Kokkos::sin(pi * x)))
+                * Kokkos::sin(Kokkos::sin(pi * y)));
+                
+    } else if (Dim == 3) {
+        T x = x_vec[0];
+        T y = x_vec[1];
+        T z = x_vec[2];
+
+        val = Kokkos::pow(pi, 2)
+                * (Kokkos::cos(Kokkos::sin(pi * z)) * Kokkos::sin(pi * z) * Kokkos::sin(Kokkos::sin(pi * x)) * Kokkos::sin(Kokkos::sin(pi * y))
+                + (Kokkos::cos(Kokkos::sin(pi * y)) * Kokkos::sin(pi * y) * Kokkos::sin(Kokkos::sin(pi * x))
+                + (Kokkos::cos(Kokkos::sin(pi * x)) * Kokkos::sin(pi * x)
+                + (Kokkos::pow(Kokkos::cos(pi * x), 2) + Kokkos::pow(Kokkos::cos(pi * y), 2) + Kokkos::pow(Kokkos::cos(pi * z), 2))
+                * Kokkos::sin(Kokkos::sin(pi * x)))
+                * Kokkos::sin(Kokkos::sin(pi * y)))
+                * Kokkos::sin(Kokkos::sin(pi * z)));
+    }
+
+    return val;
 }
 
-template <typename T>
+template <typename T, unsigned Dim>
 struct AnalyticSol {
-    KOKKOS_FUNCTION const T operator()(ippl::Vector<T, 1> x_vec) const {
-        const T pi     = Kokkos::numbers::pi_v<T>;
-        const T sqrt_2 = Kokkos::sqrt(2.0);
+    const T pi = Kokkos::numbers::pi_v<T>;
 
-        const T sigma = 0.05;
-        const T mu    = 0.5;
-
-        const T x = x_vec[0];
-        const T r = (x - mu);
-
-        return (-sigma / sqrt_2)
-                * ((r / (sqrt_2 * sigma)) * Kokkos::erf(r / (sqrt_2 * sigma))
-                + (1.0 / Kokkos::sqrt(pi)) * Kokkos::exp(-(r * r) / (2.0 * sigma * sigma)))
-                + (0.5 - mu) * x + 0.5 * mu;
+    KOKKOS_FUNCTION const T operator()(ippl::Vector<T, Dim> x_vec) const {
+        T val = 1.0;
+        for (unsigned d = 0; d < Dim; d++) {
+            val *= Kokkos::sin(Kokkos::sin(pi * x_vec[d]));
+        }
+        return val;
     }
 };
 
@@ -69,23 +102,24 @@ void testFEMSolver(const unsigned& numNodesPerDim, const T& domain_start = 0.0,
 
     ippl::FieldLayout<Dim> layout(MPI_COMM_WORLD, domain, isParallel);
     Field_t lhs(mesh, layout, numGhosts);  // left hand side (updated in the algorithm)
-    Field_t rhs(mesh, layout, numGhosts);  // right hand side (set once)
-    Field_t sol(mesh, layout, numGhosts);  // right hand side (set once)
+    Field_t rhs(mesh, layout, numGhosts);  // left hand side (updated in the algorithm)
+
+    Field_t analytical(mesh, layout, numGhosts);  // right hand side (set once)
+    auto view_analytical = analytical.getView();
 
     // Define boundary conditions
     BConds_t bcField;
     for (unsigned int i = 0; i < 2 * Dim; ++i) {
-        bcField[i] = std::make_shared<ippl::ZeroFace<Field_t>>(i);
+        bcField[i] = std::make_shared<ippl::PeriodicFace<Field_t>>(i);
     }
     lhs.setFieldBC(bcField);
     rhs.setFieldBC(bcField);
 
     // set rhs
     auto view_rhs = rhs.getView();
-    auto view_sol = sol.getView();
     auto ldom     = layout.getLocalNDIndex();
 
-    AnalyticSol<T> analytic;
+    AnalyticSol<T, Dim> analytic;
 
     using index_array_type = typename ippl::RangePolicy<Dim>::index_array_type;
     ippl::parallel_for(
@@ -94,10 +128,11 @@ void testFEMSolver(const unsigned& numNodesPerDim, const T& domain_start = 0.0,
             for (unsigned d = 0; d < Dim; ++d) {
                 iVec[d] += ldom[d].first();
             }
+
             const ippl::Vector<T, Dim> x = (iVec)*cellSpacing + origin;
 
-            apply(view_rhs, args) = gaussian1D<T>(x);
-            apply(view_sol, args) = analytic(x);
+            apply(view_rhs, args) = sinusoidalRHSFunction<T, Dim>(x);
+            apply(view_analytical, args) = analytic(x);
         });
 
     IpplTimings::stopTimer(initTimer);
@@ -109,7 +144,6 @@ void testFEMSolver(const unsigned& numNodesPerDim, const T& domain_start = 0.0,
     ippl::ParameterList params;
     params.add("tolerance", 1e-13);
     params.add("max_iterations", 2000);
-
     solver.mergeParameters(params);
 
     // solve the problem
@@ -122,8 +156,8 @@ void testFEMSolver(const unsigned& numNodesPerDim, const T& domain_start = 0.0,
     // Compute the error
     const T relError = solver.getL2Error(analytic);
 
-    lhs = lhs - sol;
-    const T normError = norm(lhs) / norm(sol);
+    lhs = lhs - analytical;
+    T normError = norm(lhs) / norm(analytical);
 
     m << std::setw(10) << numNodesPerDim;
     m << std::setw(25) << std::setprecision(16) << cellSpacing[0];
@@ -143,6 +177,14 @@ int main(int argc, char* argv[]) {
 
         using T = double;
 
+        unsigned dim = 3;
+
+        if (argc > 1 && std::atoi(argv[1]) == 1) {
+            dim = 1;
+        } else if (argc > 1 && std::atoi(argv[1]) == 2) {
+            dim = 2;
+        }
+
         // start the timer
         static IpplTimings::TimerRef allTimer = IpplTimings::getTimer("allTimer");
         IpplTimings::startTimer(allTimer);
@@ -155,8 +197,21 @@ int main(int argc, char* argv[]) {
         msg << std::setw(15) << "Iterations";
         msg << endl;
 
-        for (unsigned n = 1 << 2; n <= 1 << 9; n = n << 1) {
-            testFEMSolver<T, 1>(n, 0.0, 1.0);
+        if (dim == 1) {
+            // 1D Sinusoidal
+            for (unsigned n = 1 << 3; n <= 1 << 10; n = n << 1) {
+                testFEMSolver<T, 1>(n, -1.0, 1.0);
+            }
+        } else if (dim == 2) {
+            // 2D Sinusoidal
+            for (unsigned n = 1 << 3; n <= 1 << 10; n = n << 1) {
+                testFEMSolver<T, 2>(n, -1.0, 1.0);
+            }
+        } else {
+            // 3D Sinusoidal
+            for (unsigned n = 1 << 3; n <= 1 << 9; n = n << 1) {
+                testFEMSolver<T, 3>(n, -1.0, 1.0);
+            }
         }
 
         // stop the timer
