+++ test/solver/fem/TestScaling_PeriodicBC_sinsin.cpp
@@ -1,16 +1,28 @@
-// Tests the FEM Poisson solver by solving the problem:
+// Periodic BCs scaling test
 //
-// -Laplacian(u) = pi^2 * sin(pi * x), x in [-1,1]
-// u(-1) = u(1) = 0
+// Run the solve with periodic BCs 5 times,
+// with problem size as user input.
+// This is the 3D case.
 //
-// Exact solution is u(x) = sin(pi * x)
+// Uses the sinusoidal as source function
+// (see TestPeriodicBC_sinsin.cpp)
 //
 // Usage:
-//    ./TestFEMPoissonSolver <dim> --info 5
+//  ./TestScaling_PeriodicBC_sinsin <pt> --info 5
 
 #include "Ippl.h"
 
-#include "Meshes/Centering.h"
+#include <Kokkos_MathematicalConstants.hpp>
+#include <Kokkos_MathematicalFunctions.hpp>
+#include <cstdlib>
+#include <iostream>
+#include <string>
+#include <typeinfo>
+
+#include "Utility/Inform.h"
+#include "Utility/IpplTimings.h"
+
+#include "PoissonSolvers/PoissonCG.h"
 #include "PoissonSolvers/FEMPoissonSolver.h"
 
 template <typename T, unsigned Dim>
@@ -18,11 +30,38 @@ KOKKOS_INLINE_FUNCTION T sinusoidalRHSFunction(ippl::Vector<T, Dim> x_vec) {
     const T pi = Kokkos::numbers::pi_v<T>;
 
     T val = 1.0;
-    for (unsigned d = 0; d < Dim; d++) {
-        val *= Kokkos::sin(pi * x_vec[d]);
+    if (Dim == 1) {
+        T x = x_vec[0];
+
+        val = Kokkos::pow(pi, 2) * ((Kokkos::cos(Kokkos::sin(pi * x)) * Kokkos::sin(pi * x))
+                + (Kokkos::pow(Kokkos::cos(pi * x), 2) * Kokkos::sin(Kokkos::sin(pi * x))));
+
+    } else if (Dim == 2) {
+        T x = x_vec[0];
+        T y = x_vec[1];
+
+        val = Kokkos::pow(pi, 2)
+                * (Kokkos::cos(Kokkos::sin(pi * y)) * Kokkos::sin(pi * y) * Kokkos::sin(Kokkos::sin(pi * x))
+                + (Kokkos::cos(Kokkos::sin(pi * x)) * Kokkos::sin(pi * x)
+                + (Kokkos::pow(Kokkos::cos(pi * x), 2) + Kokkos::pow(Kokkos::cos(pi * y), 2)) * Kokkos::sin(Kokkos::sin(pi * x)))
+                * Kokkos::sin(Kokkos::sin(pi * y)));
+                
+    } else if (Dim == 3) {
+        T x = x_vec[0];
+        T y = x_vec[1];
+        T z = x_vec[2];
+
+        val = Kokkos::pow(pi, 2)
+                * (Kokkos::cos(Kokkos::sin(pi * z)) * Kokkos::sin(pi * z) * Kokkos::sin(Kokkos::sin(pi * x)) * Kokkos::sin(Kokkos::sin(pi * y))
+                + (Kokkos::cos(Kokkos::sin(pi * y)) * Kokkos::sin(pi * y) * Kokkos::sin(Kokkos::sin(pi * x))
+                + (Kokkos::cos(Kokkos::sin(pi * x)) * Kokkos::sin(pi * x)
+                + (Kokkos::pow(Kokkos::cos(pi * x), 2) + Kokkos::pow(Kokkos::cos(pi * y), 2) + Kokkos::pow(Kokkos::cos(pi * z), 2))
+                * Kokkos::sin(Kokkos::sin(pi * x)))
+                * Kokkos::sin(Kokkos::sin(pi * y)))
+                * Kokkos::sin(Kokkos::sin(pi * z)));
     }
 
-    return Dim * pi * pi * val;
+    return val;
 }
 
 template <typename T, unsigned Dim>
@@ -32,7 +71,7 @@ struct AnalyticSol {
     KOKKOS_FUNCTION const T operator()(ippl::Vector<T, Dim> x_vec) const {
         T val = 1.0;
         for (unsigned d = 0; d < Dim; d++) {
-            val *= Kokkos::sin(pi * x_vec[d]);
+            val *= Kokkos::sin(Kokkos::sin(pi * x_vec[d]));
         }
         return val;
     }
@@ -68,12 +107,15 @@ void testFEMSolver(const unsigned& numNodesPerDim, const T& domain_start = 0.0,
 
     ippl::FieldLayout<Dim> layout(MPI_COMM_WORLD, domain, isParallel);
     Field_t lhs(mesh, layout, numGhosts);  // left hand side (updated in the algorithm)
-    Field_t rhs(mesh, layout, numGhosts);  // right hand side (set once)
+    Field_t rhs(mesh, layout, numGhosts);  // left hand side (updated in the algorithm)
+
+    Field_t analytical(mesh, layout, numGhosts);  // right hand side (set once)
+    auto view_analytical = analytical.getView();
 
     // Define boundary conditions
     BConds_t bcField;
     for (unsigned int i = 0; i < 2 * Dim; ++i) {
-        bcField[i] = std::make_shared<ippl::ZeroFace<Field_t>>(i);
+        bcField[i] = std::make_shared<ippl::PeriodicFace<Field_t>>(i);
     }
     lhs.setFieldBC(bcField);
     rhs.setFieldBC(bcField);
@@ -81,6 +123,8 @@ void testFEMSolver(const unsigned& numNodesPerDim, const T& domain_start = 0.0,
     // set rhs
     auto view_rhs = rhs.getView();
     auto ldom     = layout.getLocalNDIndex();
+
+    AnalyticSol<T, Dim> analytic;
 
     using index_array_type = typename ippl::RangePolicy<Dim>::index_array_type;
     ippl::parallel_for(
@@ -93,6 +137,7 @@ void testFEMSolver(const unsigned& numNodesPerDim, const T& domain_start = 0.0,
             const ippl::Vector<T, Dim> x = (iVec)*cellSpacing + origin;
 
             apply(view_rhs, args) = sinusoidalRHSFunction<T, Dim>(x);
+            apply(view_analytical, args) = analytic(x);
         });
 
     IpplTimings::stopTimer(initTimer);
@@ -114,12 +159,15 @@ void testFEMSolver(const unsigned& numNodesPerDim, const T& domain_start = 0.0,
     IpplTimings::startTimer(errorTimer);
 
     // Compute the error
-    AnalyticSol<T, Dim> analytic;
     const T relError = solver.getL2Error(analytic);
+
+    lhs = lhs - analytical;
+    T normError = norm(lhs) / norm(analytical);
 
     m << std::setw(10) << numNodesPerDim;
     m << std::setw(25) << std::setprecision(16) << cellSpacing[0];
     m << std::setw(25) << std::setprecision(16) << relError;
+    m << std::setw(25) << std::setprecision(16) << normError;
     m << std::setw(25) << std::setprecision(16) << solver.getResidue();
     m << std::setw(15) << std::setprecision(16) << solver.getIterationCount();
     m << endl;
@@ -134,14 +182,6 @@ int main(int argc, char* argv[]) {
 
         using T = double;
 
-        unsigned dim = 3;
-
-        if (argc > 1 && std::atoi(argv[1]) == 1) {
-            dim = 1;
-        } else if (argc > 1 && std::atoi(argv[1]) == 2) {
-            dim = 2;
-        }
-
         // start the timer
         static IpplTimings::TimerRef allTimer = IpplTimings::getTimer("allTimer");
         IpplTimings::startTimer(allTimer);
@@ -153,21 +193,10 @@ int main(int argc, char* argv[]) {
         msg << std::setw(15) << "Iterations";
         msg << endl;
 
-        if (dim == 1) {
-            // 1D Sinusoidal
-            for (unsigned n = 1 << 3; n <= 1 << 10; n = n << 1) {
-                testFEMSolver<T, 1>(n, -1.0, 1.0);
-            }
-        } else if (dim == 2) {
-            // 2D Sinusoidal
-            for (unsigned n = 1 << 3; n <= 1 << 10; n = n << 1) {
-                testFEMSolver<T, 2>(n, -1.0, 1.0);
-            }
-        } else {
-            // 3D Sinusoidal
-            for (unsigned n = 1 << 3; n <= 1 << 9; n = n << 1) {
-                testFEMSolver<T, 3>(n, -1.0, 1.0);
-            }
+        unsigned pt = std::atoi(argv[1]);
+
+        for (unsigned n = 0; n < 5; ++n) {
+            testFEMSolver<T, 3>(pt, -1.0, 1.0);
         }
 
         // stop the timer
